refactor(api): 同居型テスト 26 件を src/__tests__/ へ集約しテスト配置を統一
- AGENTS.md のテスト配置規約（server/api は集中型）に従い、
  src/{lib,services,utils,agents,content} 直下の *.test.ts を
  src/__tests__/ のミラー位置へ移動。相対 import を追従更新。
- composeSessionRunLocale.test.ts の bun:test import を vitest に修正
  （vitest 実行時に suite ごと失敗していた既存問題。CI は api の
  vitest を実行していないため未検出だった）

検証: server/api で tsc --noEmit クリーン、vitest 全 1424 テストがパス。
