+++ server/api/src/__tests__/agents/core/composeModelConfig.test.ts
@@ -2,8 +2,8 @@
  * `getComposeModelIdsForGraph` unit tests — BYOK validation inputs (#953).
  */
 import { describe, expect, it } from "vitest";
-import { getComposeModelIdsForGraph } from "./composeModelConfig.js";
-import { WIKI_MAINTENANCE_GRAPH_ID } from "../graphs/wikiMaintenance/index.js";
+import { getComposeModelIdsForGraph } from "../../../agents/core/composeModelConfig.js";
+import { WIKI_MAINTENANCE_GRAPH_ID } from "../../../agents/graphs/wikiMaintenance/index.js";
 
 describe("getComposeModelIdsForGraph", () => {
   it("returns no model ids for wiki-maintenance (lint-only graph)", () => {
