+++ server/api/src/__tests__/lib/extAuth.test.ts
@@ -11,7 +11,7 @@ import {
   issueExtensionToken,
   verifyExtensionToken,
   type ExtensionTokenPayload,
-} from "./extAuth.js";
+} from "../../lib/extAuth.js";
 
 /** base64url encode (no padding) - mirrors extAuth internal logic for test */
 function sha256Base64url(str: string): string {
