+++ server/api/src/__tests__/services/lintEngine/rules/orphan.test.ts
@@ -3,9 +3,9 @@
  * Tests for the orphan rule's row → finding mapping.
  */
 import { describe, it, expect } from "vitest";
-import { runOrphanRule } from "./orphan.js";
-import { createMockDb } from "../../../__tests__/createMockDb.js";
-import type { Database } from "../../../types/index.js";
+import { runOrphanRule } from "../../../../services/lintEngine/rules/orphan.js";
+import { createMockDb } from "../../../createMockDb.js";
+import type { Database } from "../../../../types/index.js";
 
 function asDb(results: unknown[]) {
   const { db, chains } = createMockDb(results);
