+++ server/api/src/__tests__/services/notifier.test.ts
@@ -15,12 +15,12 @@ import { describe, it, expect, vi, beforeEach, afterEach } from "vitest";
 
 const mockSendEmail = vi.fn();
 
-vi.mock("./emailService.js", () => ({
+vi.mock("../../services/emailService.js", () => ({
   sendEmail: mockSendEmail,
 }));
 
 // モック設定後にインポート / Import after mock setup
-const { notifyApiErrorAlert } = await import("./notifier.js");
+const { notifyApiErrorAlert } = await import("../../services/notifier.js");
 
 describe("notifyApiErrorAlert", () => {
   const originalEnv = process.env;
