+++ server/api/src/__tests__/services/syncAiModelsFilters.test.ts
@@ -8,7 +8,7 @@ import {
   isSonnetModel,
   assignTier,
   parseAllowlist,
-} from "./syncAiModelsFilters.js";
+} from "../../services/syncAiModelsFilters.js";
 
 describe("isTextChatModel", () => {
   it("openai: gpt-4 は true", () => {
